security2 - implement. interface AuthenticationEntryPoint
To customize the response for a failed authentication, we can implement an AuthenticationEntryPoint.
Its method commence() receives the HttpServletRequest, the HttpServletResponse, and the AuthenticationException
that caused the authentication to fail.

It’s a little bit ambiguous that the name of the AuthenticationEntryPoint interface doesn’t reflect its usage upon
authentication failure. In the Spring Security architecture, this is used directly by a component called ExceptionTranslationManager.